requirements.txt
Django==4.2.7
djangorestframework==3.15.2
python-decouple==3.8
psycopg2-binary
Pillow>=11.0.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0
razorpay==2.0.0
celery==5.4.0
redis==5.1.0
requests==2.32.3
ffmpeg-python==0.2.0
gunicorn==21.2.0
django-celery-beat==2.6.0
flower==2.0.1
psutil
django-storages
django-filter
django-redis
boto3
django-ratelimit
python-magic-bin
whitenoise==6.5.0
yt-dlp